requirements.txt
# ==========================================
# 🧠 ASANA SIMULATION REQUIREMENTS (Python 3.9)
# ==========================================

# Core Data Utilities
pandas==1.5.3
numpy==1.24.4
faker==24.11.0
tqdm==4.66.2

# Configuration
python-dotenv==1.0.1
typing-extensions==4.8.0

# LLM Integration
openai==1.10.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
